refactor(cli): resolve OpenClaw workspace once
